remove state from board page and shift to react query
